requirements.txt
# python==3.7.7
uvicorn
fastapi
sqlalchemy
mysqlclient
# async-exit-stack
# async-generator

ujson
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
# python-socketio
#requests


# flask==1.1.1
# flask-sqlalchemy==2.4.1
# sqlalchemy
# mysqlclient==1.4.6
# flask-login==0.4.1
# gunicorn==20.0.4
# gevent==1.4.0
# flask-sse==0.2.1
# Flask-APScheduler==1.11.0
